requirements.txt
Django==1.3
Jinja2==2.5.5
Djinja
boto

unicode-slugify
django-taggit
django-debug-toolbar
-e git+https://github.com/jessemiller/HamlPy#egg=hamlpy
django-mediagenerator